requirements.txt
# Auto-generated from pyproject.toml for Railway deployment
# Core
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy>=2.0

# Email parsing
mail-parser>=3.15
beautifulsoup4>=4.12
lxml>=5.0

# Gmail API
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0

# AI/LLM
anthropic>=0.25

# Substack (unofficial)
python-substack @ git+https://github.com/ma2za/python-substack.git

# Web framework
fastapi>=0.110
uvicorn>=0.29

# Email sending
resend>=1.0

# Scheduling
apscheduler>=3.10

# Security
itsdangerous>=2.1
cryptography>=42.0

# HTTP client
httpx>=0.27

# Utilities
python-dotenv>=0.21.0
tenacity>=8.2
structlog>=24.1

# Zillow enrichment
apify-client>=1.6.0
usaddress>=0.5.10
rapidfuzz>=3.5.0
